sv_timestamp_logger: output stream unique hex int instead of identifier
Currently, the sv_timestamp_logger output each SV stream with string its
identifier. As there is no specification about the format of the stream
identifier, it is not possible to determine easily the stream number.
In particular, when post-processing the data we want to make the
analysis on a stream number rather than on a stream identifier.

To fix this, this commit changes the output format of the
sv_timestamp_logger to use a unique hexadecimal integer for each stream
instead of its string identifier.
Ex:
svID0000 --> 0x0000
svID0001 --> 0x0001
...
svID0007 --> 0x0007

The unique hexadecimal integer is set depending on the first SV stream
received.

Signed-off-by: Paul Le Guen de Kerneizon <paul.leguendekerneizon@savoirfairelinux.com>